[feature][CAR-22] Create filtered cars endpoint
