+++ backend/src/modules/cars/cars.module.ts
@@ -2,12 +2,12 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { CarEntity } from '../../shared/entities/car.entity';
 import { ImageEntity } from '../../shared/entities/image.entity';
-import { RecommendedCarsController } from './recommended-cars.controller';
-import { RecommendedCarsService } from './recommended-cars.service';
+import { CarsController } from './cars.controller';
+import { CarsService } from './cars.service';
 
 @Module({
     imports: [TypeOrmModule.forFeature([CarEntity, ImageEntity])],
-    controllers: [RecommendedCarsController],
-    providers: [RecommendedCarsService],
+    controllers: [CarsController],
+    providers: [CarsService],
 })
-export class RecommendedCarsModule {}
+export class CarsModule {}
